refactor(web): src/app을 루트 app으로 이동 및 경로 설정 수정
