+++ apps/web/app/api/debug/sentry/route.ts
