+++ apps/web/app/debug/sentry/page.tsx
