+++ apps/web/app/favicon.ico
