+++ apps/web/app/layout.tsx
@@ -1,11 +1,11 @@
 import type { Metadata } from "next";
 import localFont from "next/font/local";
-import Providers from "@/shared/providers";
+import Providers from "../src/shared/providers";
 import "@nugudi/themes/themes.css";
 import "@nugudi/nugudi-ui/style.css";
 
 const Pretendard = localFont({
-  src: "../../node_modules/pretendard/dist/web/variable/woff2/PretendardVariable.woff2",
+  src: "../node_modules/pretendard/dist/web/variable/woff2/PretendardVariable.woff2",
   display: "swap",
   weight: "45 920",
   adjustFontFallback: false,
