+++ apps/web/app/manifest.ts
