+++ apps/web/src/shared/providers/index.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import type { PropsWithChildren } from "react";
-import { ThemeScript } from "@/shared/providers/scripts/theme-script";
+import { ThemeScript } from "@/src/shared/providers/scripts/theme-script";
 import { RegisterServiceWorkerClient } from "../configs/pwa/register-service-worker-client";
 import QueryProvider from "./modules/query-provider";
