+++ apps/web/src/shared/providers/modules/query-provider.tsx
@@ -3,7 +3,7 @@
 import { QueryClientProvider } from "@tanstack/react-query";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
 import type { PropsWithChildren } from "react";
-import getQueryClient from "@/shared/configs/tanstack-query/get-query-client";
+import getQueryClient from "@/src/shared/configs/tanstack-query/get-query-client";
 
 function QueryProvider({ children }: PropsWithChildren) {
   const queryClient = getQueryClient();
